small fix with celkrokow as int and now MaxStep is updating after every change in profile

+++ InzynieriaAplikacja/Models/User.cs
@@ -16,7 +16,7 @@ public class User
     public string Password { get; set; }
     public float Waga { get; set; }
     public float Wzrost { get; set; }
-    public float CelKrokow { get; set; }
+    public int CelKrokow { get; set; }
 
 
     [ForeignKey(typeof(Statistic))]
